backend/requirements.txt
django == 5.2.4
djangorestframework == 3.16.0
 djangorestframework-simplejwt == 5.5.0
celery == 5.5.3
redis == 6.2.0
PyMuPDF == 1.26.3
python-docx == 1.2.0
django-cors-headers == 4.7.0
sendgrid == 6.12.4